fix swagger

+++ Chronolibris.Application/Handlers/Books/DeleteBookFileHandler.cs
@@ -21,21 +21,27 @@ namespace Chronolibris.Application.Handlers.Books
         public async Task<Unit> Handle(DeleteBookFileCommand request, CancellationToken cancellationToken)
         {
             var bookFile = await _unitOfWork.BookFiles.GetByIdAsync(request.BookFileId, cancellationToken);
-            if (bookFile == null)
+            if (bookFile == null || bookFile.StatusId==BookFileStatuses.DELETED)
                 return Unit.Value;
 
-            if (bookFile.StatusId != BookFileStatuses.ARCHIVE && bookFile.IsReadable == true)
+            if (bookFile.StatusId != BookFileStatuses.ARCHIVE && bookFile.StatusId!=BookFileStatuses.FAILED)
             {
                 bookFile.StatusId = BookFileStatuses.ARCHIVE;
                 bookFile.HiddenAt = DateTime.UtcNow;
+                bookFile.HiddenBy = request.UserId;
                 _unitOfWork.BookFiles.Update(bookFile);
                 await _unitOfWork.SaveChangesAsync(cancellationToken);
             }
             else
             {
-                _unitOfWork.BookFiles.Delete(bookFile);
-                await _unitOfWork.SaveChangesAsync(cancellationToken);
-                await _bookStorage.DeleteBookDataAsync(bookFile.Id.ToString(), cancellationToken);
+                bookFile.StatusId=BookFileStatuses.DELETED;
+                bookFile.DeletedAt = DateTime.UtcNow;
+                bookFile.DeletedBy = request.UserId;
+                //_unitOfWork.BookFiles.Delete(bookFile);
+                //await _unitOfWork.SaveChangesAsync(cancellationToken);
+                _unitOfWork.BookFiles.Update(bookFile);
+                var res = await _unitOfWork.SaveChangesAsync(cancellationToken);
+                if(res>0) await _bookStorage.DeleteBookDataAsync(bookFile.Id.ToString(), cancellationToken);
             }
 
             return Unit.Value;
